Introduce ArgProvider
