+++ devito/tools/abc.py
@@ -1,11 +1,11 @@
-from abc import ABC
+import abc
 from hashlib import sha1
 
 
-__all__ = ['Tag', 'Signer', 'Pickable']
+__all__ = ['Tag', 'ArgProvider', 'Signer', 'Pickable']
 
 
-class Tag(ABC):
+class Tag(abc.ABC):
 
     """
     An abstract class to define categories of object decorators.
@@ -35,6 +35,44 @@ class Tag(ABC):
         return ret
 
     __repr__ = __str__
+
+
+class ArgProvider(object):
+
+    """
+    A base class for types that can provide runtime values for dynamically
+    executed (JIT-compiled) code.
+    """
+
+    @abc.abstractproperty
+    def _arg_names(self):
+        raise NotImplementedError('%s does not provide any default argument names' %
+                                  self.__class__)
+
+    @abc.abstractmethod
+    def _arg_defaults(self):
+        """
+        Returns a map of default argument values defined by this type.
+        """
+        raise NotImplementedError('%s does not provide any default arguments' %
+                                  self.__class__)
+
+    @abc.abstractmethod
+    def _arg_values(self, **kwargs):
+        """
+        Returns a map of argument values after evaluating user input.
+
+        :param kwargs: Dictionary of user-provided argument overrides.
+        """
+        raise NotImplementedError('%s does not provide argument value derivation' %
+                                  self.__class__)
+
+    @abc.abstractmethod
+    def _arg_check(self, *args, **kwargs):
+        """
+        Raises an exception if an argument value is illegal.
+        """
+        raise NotImplementedError('%s does not support argument check' % self.__class__)
 
 
 class Signer(object):
